+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## In development
+### Bugfixes
+- `ExecuteAsUser` and `ExecuteForUsers` didn't execute locally if the id of the current user was passed in as recipient.
+
+
 ## 1.0.1
 ### New features
 - Added support for game systems